+++ packages/openfortkit/src/components/Common/PoweredByFooter/index.tsx
@@ -1,7 +1,7 @@
 import { motion } from "framer-motion";
 import Logos from "../../../assets/logos";
 import styled from "../../../styles/styled";
-import { useOpenfortKit } from "../../OpenfortKit";
+import { useOpenfortKit } from '../../OpenfortKit/useOpenfortKit';
 
 const PoweredByFooter = ({ showDisclaimer }: { showDisclaimer?: boolean }) => {
   const { options } = useOpenfortKit();
