+++ packages/openfortkit/src/components/ConnectModal/index.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useAccount } from 'wagmi';
 import { CustomTheme, Languages, Mode, Theme } from '../../types';
 import Modal from '../Common/Modal';
-import { routes, useOpenfortKit } from '../OpenfortKit';
+import { useOpenfortKit } from '../OpenfortKit/useOpenfortKit';
 
 import About from '../Pages/About';
 import Connectors from '../Pages/Connectors';
@@ -17,8 +17,9 @@ import { OAuthProvider } from '@openfort/openfort-js';
 import { ValueOf } from 'viem/_types/types/utils';
 import { getAppName } from '../../defaultConfig';
 import { useChainIsSupported } from '../../hooks/useChainIsSupported';
-import { useOpenfort } from '../../openfort/OpenfortProvider';
+import { useOpenfort } from '../../openfort/useOpenfort';
 import { ConnectKitThemeProvider } from '../ConnectKitThemeProvider/ConnectKitThemeProvider';
+import { routes } from '../OpenfortKit/types';
 import EmailLogin from '../Pages/EmailLogin';
 import EmailSignup from '../Pages/EmailSignup';
 import EmailVerification from '../Pages/EmailVerification';
