+++ packages/openfortkit/src/components/Pages/EmailLogin/index.tsx
@@ -1,18 +1,16 @@
-import React, { useEffect } from "react";
-import { useConnect } from "wagmi";
-import { useOpenfort } from "../../../openfort/OpenfortProvider";
-import Button from "../../Common/Button";
-import Input from "../../Common/Input";
-import { InnerContainer, ModalBody, PageContent } from "../../Common/Modal/styles";
-import { routes, useOpenfortKit } from "../../OpenfortKit";
-import { RecoveryMethod } from "@openfort/openfort-js";
-import { OrDivider } from "../../Common/Modal";
-import FitText from "../../Common/FitText";
 import { AnimatePresence, Variants, motion } from 'framer-motion';
-import { AuthIcon } from "../../../assets/icons";
-import { TextContainer } from "../../ConnectButton/styles";
+import React from "react";
+import { useOpenfort } from '../../../openfort/useOpenfort';
 import { isPlayerVerified } from "../../../utils";
+import Button from "../../Common/Button";
 import { TextLinkButton } from "../../Common/Button/styles";
+import FitText from "../../Common/FitText";
+import Input from "../../Common/Input";
+import { OrDivider } from "../../Common/Modal";
+import { ModalBody, PageContent } from "../../Common/Modal/styles";
+import { TextContainer } from "../../ConnectButton/styles";
+import { useOpenfortKit } from '../../OpenfortKit/useOpenfortKit';
+import { routes } from '../../OpenfortKit/types';
 
 // TODO: Localize
