+++ packages/openfortkit/src/components/Pages/EmailSignup/index.tsx
@@ -1,13 +1,12 @@
-import React, { useEffect } from "react";
-import { useConnect } from "wagmi";
-import { useOpenfort } from "../../../openfort/OpenfortProvider";
+import React from "react";
+import { useOpenfort } from '../../../openfort/useOpenfort';
 import Button from "../../Common/Button";
-import Input from "../../Common/Input";
-import { ModalBody, PageContent } from "../../Common/Modal/styles";
-import { routes, useOpenfortKit } from "../../OpenfortKit";
-import { RecoveryMethod } from "@openfort/openfort-js";
-import { OrDivider } from "../../Common/Modal";
 import FitText from "../../Common/FitText";
+import Input from "../../Common/Input";
+import { OrDivider } from "../../Common/Modal";
+import { ModalBody, PageContent } from "../../Common/Modal/styles";
+import { useOpenfortKit } from '../../OpenfortKit/useOpenfortKit';
+import { routes } from "../../OpenfortKit/types";
 
 // TODO: Localize
