+++ packages/openfortkit/src/components/Pages/Onboarding/index.tsx
@@ -22,7 +22,7 @@ import wave from '../../../assets/wave';
 
 import Button from '../../Common/Button';
 import useLocales from '../../../hooks/useLocales';
-import { useOpenfortKit } from '../../OpenfortKit';
+import { useOpenfortKit } from '../../OpenfortKit/useOpenfortKit';
 
 const Introduction: React.FC = () => {
   const context = useOpenfortKit();
