+++ packages/openfortkit/src/hooks/openfort/useUser.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { useOpenfort } from "../../openfort/OpenfortProvider";
+import { useOpenfort } from '../../openfort/useOpenfort';
 
 export function useUser() {
   const [accessToken, setAccessToken] = useState<string | null>(null);
