+++ packages/openfortkit/src/hooks/useLocales.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import Logos from '../assets/logos';
 
-import { useOpenfortKit } from '../components/OpenfortKit';
+import { useOpenfortKit } from '../components/OpenfortKit/useOpenfortKit';
 
 import { getLocale } from '../localizations';
