+++ packages/openfortkit/src/openfort/context.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { ContextValue } from './OpenfortProvider';
+
+export const Context = createContext<ContextValue | null>(null);
