+++ packages/openfortkit/src/types.ts
@@ -20,5 +20,5 @@ export type All = {
   lang?: Languages;
 };
 
-export type { ConnectKitOptions } from './components/OpenfortKit';
+export type { ConnectKitOptions } from './components/OpenfortKit/types';
 export type { CustomAvatarProps } from './components/Common/Avatar';
