+++ packages/openfortkit/src/version.ts
@@ -1,0 +1,1 @@
+export const OPENFORTKIT_VERSION = '0.0.6';
